+++ iOS/Omnom/Omnom/Model/Tip/OMNTip.h
@@ -10,16 +10,12 @@
 
 @interface OMNTip : NSObject
 
-@property (nonatomic, assign) double amount;
+@property (nonatomic, assign) long long amount;
 @property (nonatomic, assign) double percent;
-//@property (nonatomic, assign) double calculationValue;
 @property (nonatomic, assign) BOOL selected;
 
 + (instancetype)tipWithPercent:(double)percent;
 
-- (instancetype)initWithAmount:(double)amount percent:(double)percent;
-
-//- (NSString *)title;
-//- (NSString *)selectedTitle;
+- (instancetype)initWithAmount:(long long)amount percent:(double)percent;
 
 @end
